requirements.txt
librosa==0.7.0
opencv-contrib-python>=4.2.0.34
opencv-python==4.1.0.25
imageio==2.3.0
matplotlib==2.2.2
numpy==1.17.0
pandas==0.23.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.1
scikit-image==0.14.5
scikit-learn==0.19.2
scipy==1.1.0
tqdm==4.24.0
numba==0.48